refactor document and message components for improved data handling and user feedback; add loading state for documents

+++ src/app/documents/document-detail/document-detail.component.ts
@@ -1,21 +1,21 @@
-import { Component, OnInit } from '@angular/core';
-
+import { Component, OnInit, OnDestroy } from '@angular/core';
 import { Document } from '../document.model';
 import { DocumentService } from '../documents.service';
 import { ActivatedRoute, Params, Router } from '@angular/router';
 import { WindRefService } from '../../wind-ref.service';
+import { Subscription } from 'rxjs';
 
 @Component({
   selector: 'app-document-detail',
-  standalone: false,
-  
   templateUrl: './document-detail.component.html',
-  styleUrl: './document-detail.component.css'
+  styleUrl: './document-detail.component.css',
+  standalone: false
 })
-export class DocumentDetailComponent implements OnInit{
+export class DocumentDetailComponent implements OnInit, OnDestroy {
   document: Document;
   id: string;
   nativeWindow: any;
+  subscription: Subscription;
 
   constructor(private documentService: DocumentService,
               private route: ActivatedRoute,
@@ -27,20 +27,42 @@ export class DocumentDetailComponent implements OnInit{
     this.route.params.subscribe(
       (params: Params) => {
         this.id = params['id'];
-        this.document = this.documentService.getDocument(this.id);
-        this.nativeWindow = this.window.getNativeWindow();
+        
+        // Subscribe to the document service to get the documents
+        this.subscription = this.documentService.getDocuments().subscribe(
+          (documents: Document[]) => {
+            // Now we can find the document by id from the fetched documents
+            this.document = this.documentService.getDocument(this.id);
+            
+            // If document is still null, redirect to documents page
+            if (!this.document) {
+              this.router.navigate(['/documents']);
+              return;
+            }
+            
+            this.nativeWindow = this.window.getNativeWindow();
+          }
+        );
       }
     );
   }
 
-
   onView() {
-    this.nativeWindow.open(this.document.url);
+    if (this.document && this.document.url) {
+      this.nativeWindow.open(this.document.url);
+    }
   }
 
   onDelete() {
-    this.documentService.deleteDocument(this.document);
-    this.router.navigate(['/documents'], {relativeTo: this.route});
- }
-
+    if (this.document) {
+      this.documentService.deleteDocument(this.document);
+      this.router.navigate(['/documents']);
+    }
+  }
+  
+  ngOnDestroy() {
+    if (this.subscription) {
+      this.subscription.unsubscribe();
+    }
+  }
 }
